+++ src/Functional/HStringMap.php
@@ -1,6 +1,7 @@
 <?php
 namespace Haystack\Functional;
 
+use Haystack\HArray;
 use Haystack\HString;
 
 class HStringMap
@@ -16,16 +17,57 @@ class HStringMap
         $this->hString = $hString;
     }
 
-    public function map(callable $func)
+    /**
+     * @param callable $func
+     * @param array    $variadicList
+     * @return HString
+     */
+    public function map(callable $func, $variadicList = [])
     {
-        $newString = new HString($this->hString);
-
-        $size = $this->hString->count();
-        for ($i = 0; $i < $size; $i++) {
-            $newString[$i] = $func($this->hString[$i]);
+        if (empty($variadicList)) {
+            return $this->hString->toHArray()->map($func)->toHString();
         }
 
-        return $newString;
+        $arrays = [$this->convertHStringToArrayOfChars($this->hString)];
+
+        foreach ($variadicList as $item) {
+            if (is_string($item)) {
+                $item = new HString($item);
+            }
+
+            if ($item instanceof HString) {
+                $item = $this->convertHStringToArrayOfChars($item);
+            }
+
+            if ($item instanceof HArray) {
+                $item = $item->toArray();
+            }
+
+            if (is_array($item)) {
+                $arrays[] = $item;
+            } else {
+                throw new \InvalidArgumentException("{$item} cannot be mapped");
+            }
+        }
+
+        $result = call_user_func_array('array_map', array_merge([$func], $arrays));
+
+        return (new HArray($result))->toHString();
+    }
+
+    /**
+     * @param HString $hString
+     * @return array
+     */
+    private function convertHStringToArrayOfChars(HString $hString)
+    {
+        $arr = [];
+
+        foreach ($hString as $char) {
+            $arr[] = $char;
+        }
+
+        return $arr;
     }
 
 }