+++ src/HString.php
@@ -232,10 +232,25 @@ class HString implements \Iterator, \ArrayAccess, \Serializable, \Countable, Con
     }
 
     /**
+     * Converts a string into an array. Assumes a delimiter of "" to return an array of chars.
+     *
+     * @return array
+     */
+    public function toArray()
+    {
+        if (empty($this->str)) {
+            return [];
+        }
+
+        return (new StringToArray($this->str))
+            ->stringToArray();
+    }
+
+    /**
      * Alias to PHP function `explode`
      *
      * @param string $delim
-     * @param int    $limit
+     * @param null|int    $limit
      * @return HArray
      * @throws \InvalidArgumentException
      */
@@ -339,7 +354,13 @@ class HString implements \Iterator, \ArrayAccess, \Serializable, \Countable, Con
     public function map(callable $func)
     {
         $answer = new HStringMap($this);
-        return new static($answer->map($func));
+        $argArrays = array_slice(func_get_args(), 1); // remove `$func`
+
+        if (empty($argArrays)) {
+            return new static($answer->map($func));
+        }
+
+        return new static($answer->map($func, $argArrays));
     }
 
     /**